sf_to_pg/requirements.txt
openai
python-dotenv
psycopg2-binary
pandas
sqlalchemy
snowflake-connector-python[pandas]
snowflake-snowpark-python
